Specify atol instead precision [no ci]

+++ src/cpp/bindings/interfaces/DiagonalizerInterface.py.cpp
@@ -17,11 +17,11 @@ static void declare_diagonalizer_interface(nb::module_ &m, std::string const &ty
     nb::class_<DiagonalizerInterface<T>> pyclass(m, pylass_name.c_str());
     pyclass
         .def("eigh",
-             nb::overload_cast<const Eigen::SparseMatrix<T, Eigen::RowMajor> &, int>(
+             nb::overload_cast<const Eigen::SparseMatrix<T, Eigen::RowMajor> &, double>(
                  &DiagonalizerInterface<T>::eigh, nb::const_))
         .def("eigh",
              nb::overload_cast<const Eigen::SparseMatrix<T, Eigen::RowMajor> &,
-                               std::optional<real_t>, std::optional<real_t>, int>(
+                               std::optional<real_t>, std::optional<real_t>, double>(
                  &DiagonalizerInterface<T>::eigh, nb::const_));
 }
